board input power: reduce SMBus transfer size from 32 to 16 bits
Omit the fractional part from board input power and send it from BMC FW to SMC
FW as a SMBus word write rather than a block write.

+++ lib/tenstorrent/bh_arc/cm2dm_msg.h
@@ -55,7 +55,7 @@ int32_t Dm2CmPingHandler(const uint8_t *data, uint8_t size);
 int32_t Dm2CmSendCurrentHandler(const uint8_t *data, uint8_t size);
 int32_t Dm2CmSendPowerHandler(const uint8_t *data, uint8_t size);
 int32_t GetInputCurrent(void);
-uint32_t GetInputPower(void);
+uint16_t GetInputPower(void);
 int32_t Dm2CmSendFanRPMHandler(const uint8_t *data, uint8_t size);
 
 #endif
